requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
APScheduler==3.10.4
cachetools==5.3.2